moving new scripts to 'training'
